+++ episodic/orchestration/_graph_nodes.py
@@ -1,0 +1,186 @@
+"""Ports-only graph nodes for structured generation orchestration."""
+
+import importlib
+import time
+import typing as typ
+
+from episodic.orchestration._types import _log_event
+from episodic.orchestration._usage import build_generation_result
+
+if typ.TYPE_CHECKING:
+    import collections.abc as cabc
+
+    from episodic.orchestration import _dto as dto
+    from episodic.orchestration import _protocols as protocols
+    from episodic.orchestration._graph_state import GenerationGraphState
+else:
+    dto = importlib.import_module("episodic.orchestration._dto")
+    protocols = importlib.import_module("episodic.orchestration._protocols")
+
+
+type ExecuteNodeResult = (
+    dict[str, tuple[dto.ActionExecutionResult, ...]]
+    | dict[str, dto.SuspendedWorkflowResult]
+)
+
+
+class ExecuteNodeFn(typ.Protocol):
+    """Callable protocol for async execute graph nodes."""
+
+    def __call__(
+        self, state: GenerationGraphState
+    ) -> cabc.Awaitable[ExecuteNodeResult]:
+        """Return the async execute-node update for *state*."""
+        ...
+
+
+async def _plan_node(
+    state: GenerationGraphState,
+    *,
+    planner: protocols.PlannerPort,
+) -> dict[str, dto.PlannerResult]:
+    """Validate state and invoke the planner to produce a PlannerResult."""
+    request = state.request
+    correlation_id = request.correlation_id if request is not None else None
+    _log_event(
+        "debug",
+        "generation_graph.plan_node.start",
+        correlation_id=correlation_id,
+    )
+    if request is None:
+        msg = "missing required state value: request"
+        raise ValueError(msg)
+    try:
+        planner_result = await planner.plan(request)
+    except Exception as exc:
+        _log_event(
+            "error",
+            "generation_graph.plan_node.error",
+            correlation_id=request.correlation_id,
+            error=str(exc),
+        )
+        raise
+    result = {"planner_result": planner_result}
+    _log_event(
+        "debug",
+        "generation_graph.plan_node.finish",
+        correlation_id=request.correlation_id,
+    )
+    return result
+
+
+async def _execute_single_action(
+    action: dto.PlannedAction,
+    request: dto.GenerationOrchestrationRequest,
+    *,
+    tool_executor: protocols.ToolExecutorPort,
+    selected_execution_model: str,
+) -> dto.ActionExecutionResult:
+    """Execute one planned action and emit diagnostic log events."""
+    started_at = time.monotonic()
+    action_fields = {
+        "correlation_id": request.correlation_id,
+        "action_id": action.action_id,
+        "action_kind": str(action.action_kind),
+        "model_tier": str(action.model_tier),
+        "execution_model": selected_execution_model,
+    }
+    try:
+        action_result = await tool_executor.execute(action, request)
+    except Exception as exc:
+        _log_event(
+            "error",
+            "generation_graph.execute_node.action.error",
+            **action_fields,
+            elapsed_ms=round((time.monotonic() - started_at) * 1000, 1),
+            error=str(exc),
+        )
+        raise
+    action_fields["execution_model"] = action_result.model
+    _log_event(
+        "debug",
+        "generation_graph.execute_node.action.finish",
+        **action_fields,
+        elapsed_ms=round((time.monotonic() - started_at) * 1000, 1),
+    )
+    return action_result
+
+
+async def _execute_node(
+    state: GenerationGraphState,
+    *,
+    tool_executor: protocols.ToolExecutorPort,
+) -> dict[str, tuple[dto.ActionExecutionResult, ...]]:
+    """Validate state and execute each planned action through the tool executor."""
+    request = state.request
+    correlation_id = request.correlation_id if request is not None else None
+    _log_event(
+        "debug",
+        "generation_graph.execute_node.start",
+        correlation_id=correlation_id,
+    )
+    if request is None:
+        msg = "missing required state value: request"
+        raise ValueError(msg)
+    planner_result = state.planner_result
+    if planner_result is None:
+        msg = "missing required state value: planner_result"
+        raise ValueError(msg)
+
+    # Keep tool execution ordered so the graph mirrors application-service semantics.
+    action_results = [
+        await _execute_single_action(
+            action,
+            request,
+            tool_executor=tool_executor,
+            selected_execution_model=planner_result.plan.selected_execution_model,
+        )
+        for action in planner_result.plan.steps
+    ]
+    result = {"action_results": tuple(action_results)}
+    _log_event(
+        "debug",
+        "generation_graph.execute_node.finish",
+        correlation_id=request.correlation_id,
+    )
+    return result
+
+
+def _finish_node(
+    state: GenerationGraphState,
+) -> dict[str, dto.GenerationOrchestrationResult]:
+    """Aggregate planner and action results into a GenerationOrchestrationResult."""
+    request = state.request
+    correlation_id = request.correlation_id if request is not None else None
+    _log_event(
+        "debug",
+        "generation_graph.finish_node.start",
+        correlation_id=correlation_id,
+    )
+    if request is None:
+        msg = "missing required state value: request"
+        raise ValueError(msg)
+    planner_result = state.planner_result
+    if planner_result is None:
+        msg = "missing required state value: planner_result"
+        raise ValueError(msg)
+    try:
+        orchestration_result = build_generation_result(
+            planner_result,
+            state.action_results,
+        )
+    except Exception as exc:
+        _log_event(
+            "error",
+            "generation_graph.finish_node.error",
+            correlation_id=correlation_id,
+            error=str(exc),
+        )
+        raise
+    result = {"orchestration_result": orchestration_result}
+    _log_event(
+        "debug",
+        "generation_graph.finish_node.finish",
+        correlation_id=correlation_id,
+    )
+    return result
