+++ pyproject.toml
@@ -480,65 +480,56 @@ root_packages = ["episodic"]
 default_rule_id = "ARCH001"
 
 [[tool.hecate.groups]]
-name = "composition_root"
-prefixes = ["episodic.api.runtime", "episodic.worker.runtime"]
-allowed = [
-    "application",
-    "composition_root",
-    "domain_ports",
-    "inbound_adapter",
-    "outbound_adapter",
+name = "outbound_adapter"
+prefixes = [
+    "episodic.canonical.adapters",
+    "episodic.canonical.storage",
+    "episodic.cost.storage",
+    "episodic.cost.pricing_catalogue",
+    "episodic.llm.openai_api",
+    "episodic.llm.openai_adapter",
+    "episodic.llm.openai_client",
 ]
+allowed = ["outbound_adapter", "application", "domain_ports"]
 
 [[tool.hecate.groups]]
-name = "domain_ports"
+name = "outbound_adapter"
 prefixes = [
-    "episodic.canonical.domain",
-    "episodic.canonical.constraints",
-    "episodic.canonical.ingestion",
-    "episodic.canonical.ingestion_ports",
-    "episodic.canonical.entity_protocols",
-    "episodic.canonical.history_protocols",
-    "episodic.canonical.health",
-    "episodic.canonical.idempotency",
-    "episodic.canonical.ingestion_sources",
-    "episodic.canonical.object_store",
-    "episodic.canonical.pagination",
-    "episodic.canonical.ports",
-    "episodic.canonical.reference_protocols",
-    "episodic.canonical.unit_of_work_protocols",
-    "episodic.canonical.upload_protocols",
-    "episodic.canonical.uploads",
-    "episodic.cost.ports",
-    "episodic.cost.engine",
-    "episodic.llm.ports",
-    "episodic.metrics_ports",
+    "episodic.canonical.adapters",
+    "episodic.canonical.storage",
+    "episodic.cost.storage",
+    "episodic.cost.pricing_catalogue",
+    "episodic.llm.openai_api",
+    "episodic.llm.openai_adapter",
+    "episodic.llm.openai_client",
 ]
-allowed = ["domain_ports"]
+allowed = ["outbound_adapter", "application", "domain_ports"]
 
 [[tool.hecate.groups]]
-name = "application"
+name = "outbound_adapter"
 prefixes = [
-    "episodic.canonical.services",
-    "episodic.canonical.ingestion_service",
-    "episodic.canonical.idempotency_service",
-    "episodic.canonical.source_intake_service",
-    "episodic.canonical.profile_templates",
-    "episodic.canonical.reference_documents",
-    "episodic.cost.recorder",
-    "episodic.generation",
-    "episodic.orchestration",
+    "episodic.canonical.adapters",
+    "episodic.canonical.storage",
+    "episodic.cost.storage",
+    "episodic.cost.pricing_catalogue",
+    "episodic.llm.openai_api",
+    "episodic.llm.openai_adapter",
+    "episodic.llm.openai_client",
 ]
-allowed = ["application", "domain_ports"]
+allowed = ["outbound_adapter", "application", "domain_ports"]
 
 [[tool.hecate.groups]]
-name = "inbound_adapter"
+name = "outbound_adapter"
 prefixes = [
-    "episodic.api",
-    "episodic.worker.tasks",
-    "episodic.worker.topology",
+    "episodic.canonical.adapters",
+    "episodic.canonical.storage",
+    "episodic.cost.storage",
+    "episodic.cost.pricing_catalogue",
+    "episodic.llm.openai_api",
+    "episodic.llm.openai_adapter",
+    "episodic.llm.openai_client",
 ]
-allowed = ["inbound_adapter", "application", "domain_ports"]
+allowed = ["outbound_adapter", "application", "domain_ports"]
 
 [[tool.hecate.groups]]
 name = "outbound_adapter"
